fix: patch all the things, bw is everywhere
