+++ food/export.py
@@ -11,6 +11,7 @@ import bw2data
 from bw2data.project import projects
 from frozendict import frozendict
 
+from common import brightway_patch as brightway_patch
 from common import (
     fix_unit,
     order_json,
