+++ import_ecoinvent.py
@@ -8,6 +8,7 @@ import bw2data
 import bw2io
 from bw2data.project import projects
 
+from common import brightway_patch as brightway_patch
 from common.import_ import add_missing_substances, import_simapro_csv
 
 CURRENT_FILE_DIR = os.path.dirname(os.path.realpath(__file__))
