+++ import_food.py
@@ -11,6 +11,7 @@ import bw2io
 from bw2data.project import projects
 from bw2io.strategies.generic import link_technosphere_by_activity_hash
 
+from common import brightway_patch as brightway_patch
 from common.import_ import (
     add_missing_substances,
     import_simapro_csv,
